requirements.txt
flask
supabase
bcrypt
python-dotenv